Remove existing About View Controller
* Also moves crash debugging gesture to Settings footer view
