+++ WooCommerce/Classes/ViewRelated/Dashboard/Dashboard.storyboard
@@ -109,39 +109,6 @@
             </objects>
             <point key="canvasLocation" x="1431" y="811"/>
         </scene>
-        <!--About View Controller-->
-        <scene sceneID="YPr-jf-D90">
-            <objects>
-                <viewController storyboardIdentifier="AboutViewController" id="0FF-UX-4XZ" customClass="AboutViewController" customModule="WooCommerce" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="ivz-5E-TEz">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <tableView clipsSubviews="YES" contentMode="scaleToFill" layoutMarginsFollowReadableWidth="YES" alwaysBounceVertical="YES" dataMode="prototypes" style="grouped" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="18" sectionFooterHeight="18" translatesAutoresizingMaskIntoConstraints="NO" id="ibP-Y7-8cn">
-                                <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                                <color key="backgroundColor" systemColor="groupTableViewBackgroundColor"/>
-                                <connections>
-                                    <outlet property="dataSource" destination="0FF-UX-4XZ" id="6jO-4P-3zY"/>
-                                    <outlet property="delegate" destination="0FF-UX-4XZ" id="eUa-vM-oOL"/>
-                                </connections>
-                            </tableView>
-                        </subviews>
-                        <viewLayoutGuide key="safeArea" id="BsY-oa-k6u"/>
-                        <constraints>
-                            <constraint firstItem="ibP-Y7-8cn" firstAttribute="trailing" secondItem="ivz-5E-TEz" secondAttribute="trailing" id="2ck-f3-BSs"/>
-                            <constraint firstItem="ibP-Y7-8cn" firstAttribute="bottom" secondItem="BsY-oa-k6u" secondAttribute="bottom" id="69q-a2-QAu"/>
-                            <constraint firstItem="ibP-Y7-8cn" firstAttribute="leading" secondItem="ivz-5E-TEz" secondAttribute="leading" id="QKN-1I-cNm"/>
-                            <constraint firstItem="ibP-Y7-8cn" firstAttribute="top" secondItem="BsY-oa-k6u" secondAttribute="top" id="qyY-cu-ksZ"/>
-                        </constraints>
-                    </view>
-                    <connections>
-                        <outlet property="tableView" destination="ibP-Y7-8cn" id="Q44-bo-hMp"/>
-                    </connections>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="T2j-DL-NgU" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="2265" y="2237"/>
-        </scene>
         <!--Licenses View Controller-->
         <scene sceneID="Uj1-Ej-yhV">
             <objects>
